+++ backend/chidic-app/src/main/kotlin/com/chidicapp/api/feedpostlike/FeedPostLikeController.kt
@@ -1,8 +1,9 @@
 package com.chidicapp.api.feedpostlike
 
-import com.chidiccommon.dto.FeedLikeResponse
+import com.chidicapp.api.response.FeedLikeResponse
 import com.chidiccore.auth.annotatiton.GetUserIdFromPrincipal
 import com.chidicdomain.domain.service.feedpostlike.FeedPostLikeService
+import com.chidicdomain.dto.FeedLikeDto
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.*
 
@@ -11,13 +12,13 @@ import org.springframework.web.bind.annotation.*
 class FeedPostLikeController(
     private val feedPostLikeService: FeedPostLikeService
 ) {
-    // 내가 좋아요 했는 여부 만들어야함
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
     fun likeCount(
         @PathVariable feedPostId: Long
-    ) : FeedLikeResponse{
-        return feedPostLikeService.getLikeCount(feedPostId)
+    ) : FeedLikeResponse {
+        val feedLikeDto = feedPostLikeService.getLikeCount(feedPostId)
+        return FeedPostLikeMapper.dtoToFeedLikeResponse(feedLikeDto)
     }
 
     @PostMapping
@@ -29,6 +30,7 @@ class FeedPostLikeController(
         feedPostLikeService.likeFeedPost(userId, feedPostId)
     }
 
+    // 내가 좋아요 했는 여부 만들어야함
     @DeleteMapping
     @ResponseStatus(HttpStatus.OK)
     fun unlikeFeedPost(
@@ -36,5 +38,14 @@ class FeedPostLikeController(
         @PathVariable feedPostId: Long
     ) {
         feedPostLikeService.unlikeFeedPost(userId, feedPostId)
+    }
+}
+
+object FeedPostLikeMapper {
+    fun dtoToFeedLikeResponse(feedLikeDto: FeedLikeDto): FeedLikeResponse {
+        return FeedLikeResponse(
+            postId = feedLikeDto.postId,
+            likeCount = feedLikeDto.likeCount
+        )
     }
 }
