+++ backend/chidic-common/src/main/kotlin/com/chidicapp/api/response/FeedLikeResponse.kt
@@ -1,4 +1,4 @@
-package com.chidiccommon.dto
+package com.chidicapp.api.response
 
 data class FeedLikeResponse(
     val postId: Long,
