+++ backend/chidic-domain/src/main/kotlin/com/chidicdomain/domain/service/feedpostlike/FeedPostLikeService.kt
@@ -1,9 +1,9 @@
 package com.chidicdomain.domain.service.feedpostlike
 
-import com.chidiccommon.dto.FeedLikeResponse
+import com.chidicdomain.dto.FeedLikeDto
 
 interface FeedPostLikeService {
-    fun getLikeCount(feedPostId: Long): FeedLikeResponse
+    fun getLikeCount(feedPostId: Long): FeedLikeDto
     fun likeFeedPost(userId: Long, feedPostId: Long)
     fun unlikeFeedPost(userId: Long, feedPostId: Long)
 }
