+++ backend/chidic-domain/src/main/kotlin/com/chidicdomain/dto/FeedLikeDto.kt
@@ -1,0 +1,6 @@
+package com.chidicdomain.dto
+
+data class FeedLikeDto(
+    val postId: Long,
+    val likeCount: Long
+)
